rename ReadonlyPlugin -> Minion and merge it into its own module
